backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
transformers==4.35.0
torch==2.1.0
torchvision==0.16.0
shap==0.43.0
lime==0.2.0.1
matplotlib==3.7.2
seaborn==0.12.2
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.4.2
aiofiles==23.2.1
httpx==0.25.1
face-recognition==1.3.0
deepface==0.0.79
tensorflow==2.13.0
keras==2.13.1
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1
wordcloud==1.9.2
plotly==5.17.0
streamlit==1.28.1 